Move get_hash_value function into IdempotebtStateStorage as a static method
